# Conflicts:
#	app/Http/Controllers/GoodsIssueController.php
#	app/Http/Controllers/ProjectController.php
#	database/migrations/2018_09_07_035718_create_pro_project_table.php
#	resources/views/project/create.blade.php

+++ database/migrations/2019_01_07_090947_create_trx_work_request_detail_table.php
@@ -17,8 +17,8 @@ class CreateTrxWorkRequestDetailTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('work_request_id');
             $table->date('required_date')->nullable();
-            $table->integer('quantity');
-            $table->integer('reserved')->default(0);            
+            $table->float('quantity',15,2);
+            $table->float('reserved',15,2)->default(0);            
             $table->text('description')->nullable();
             $table->integer('type');
             $table->unsignedInteger('material_id')->nullable();
